refactor: split into binary and library targets
